requirements.txt
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
anthropic==0.40.0
pandas==2.2.0
python-dotenv==1.0.1
pytest==8.1.0
pytest-mock==3.14.0
